+++ frontend/src/Context/AuthContext.js
@@ -7,27 +7,9 @@ export const AuthContext = createContext();
 // Create the AuthProvider component
 export const AuthProvider = ({ children }) => {
     // State to store authentication status
-    const [isYouTubeAuthenticated, setIsYouTubeAuthenticated] = useState(true);
+    const [isYouTubeAuthenticated, setIsYouTubeAuthenticated] = useState(false);
     const [isSpotifyAuthenticated, setIsSpotifyAuthenticated] = useState(false);
     const [YTProfile, setYTProfile] = useState(null);
-
-    useEffect(() => {
-        // Function to fetch authentication status and profile from backend
-        const fetchAuthStatus = async () => {
-          try {
-            const response = await axios.get('videos/auth-status'); 
-            const { isGoogleAuthenticated, GoogleProfile } = response.data;
-            setIsYouTubeAuthenticated(isGoogleAuthenticated);
-            if (isGoogleAuthenticated) {
-              setYTProfile(GoogleProfile);
-            }
-          } catch (error) {
-            console.error('Error fetching authentication status:', error);
-          }
-        };
-    
-        fetchAuthStatus();
-      }, []);
 
     return (
         <AuthContext.Provider
@@ -36,7 +18,8 @@ export const AuthProvider = ({ children }) => {
                 setIsYouTubeAuthenticated,
                 isSpotifyAuthenticated,
                 setIsSpotifyAuthenticated,
-                YTProfile
+                YTProfile,
+                setYTProfile
             }}
         >
             {children}
